requirements.txt
addict==2.4.0
albumentations==0.5.2
altair==4.1.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
httplib2==0.14.0
idna==2.8
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.9.0
imgaug==0.4.0
iotop==0.6
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jpeg4py==0.1.4
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.3
meson==0.53.2
mistune==0.8.4
monotonic==1.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
notebook==6.1.5
numpy==1.19.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
packaging==20.7
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.6.1
ptyprocess==0.6.0
pyarrow==2.0.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydeck==0.5.0
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retinaface-pytorch==0.0.7
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==3.0.4
ssh-import-id==5.10
streamlit==0.72.0
system-service==0.3
systemd-python==234
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.12.8
toml==0.10.2
toolz==0.11.1
torch==1.7.1
torchvision==0.8.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.1
virtualenv==20.2.2
wadllib==1.3.3
watchdog==1.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
youtube-dl==2020.3.24
zope.interface==4.7.1
